KeldyshED

The package KeldyshED.jl is a small scale Equilibrium Exact Diagonalization solver for finite fermionic models. It can compute single-particle Green's functions in the (Matsubara) imaginary time domain, on the 2-branch Keldysh contour, and on the 3-branch Konstantinov-Perel' contour.

Its main intended use is solution of auxiliary problems required by more advanced solvers such as the Hybridization Expansion Quantum Monte Carlo and the N-crossing approximations.

The initial public version of this package was a simplified Julia port of the TRIQS/atom_diag library, but has since acquired a number of unique features, e.g. calculation of Evolution operators.

Installation

KeldyshED.jl is a registered Julia package that can be installed via the following invocation

using Pkg
Pkg.add("KeldyshED")

Public API

KeldyshED — Module

Equilibrium Exact Diagonalization library that can also compute Green's functions on the Keldysh contour.

Expressions with creation/annihilation operators of fermions

Canonical operators

Monomials

Polynomial expressions

Factory functions for polynomial expressions

Hilbert spaces, states and linear operators

Ordered set of compound indices

Full Hilbert spaces and their subspaces

Quantum states

Operators acting in Hilbert spaces

Space autopartition algorithm

Exact Diagonalization solver

Core ED types and functions

Utility functions

Single-particle Green's functions

computegf()

Auxiliary types

Density matrix and evolution operators

License

KeldyshED.jl is free software: you can redistribute it and/or modify it under the terms of the GNU General Public License as published by the Free Software Foundation, either version 3 of the License, or (at your option) any later version.

KeldyshED.jl is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU General Public License for more details.

You should have received a copy of the GNU General Public License along with KeldyshED.jl (in the file LICENSE in this directory). If not, see http://www.gnu.org/licenses/.